Hide show-unread when empty and move anchor to it
Even though unread-badge hides itself on empty value, show-unread takes
up some space and shows on the page causing nearby elements to move.
This change makes show-unread to hide completely when it has an empty
value.

Additionally the anchor link that was previously in unread-badge is now
moved to show-unread so that unread-badge can be included in the link.